+++ react/package.json
@@ -1,5 +1,6 @@
 {
   "name": "obu-recruiting",
+  "proxy": "http://127.0.0.1:6000",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
