+++ react/src/css/FileUpload.css
@@ -1,0 +1,79 @@
+/* General form styling */
+form {
+  max-width: 800px;
+  margin: 0 auto;
+  padding: 20px;
+  background-color: #eaf8e2; /* Light green background */
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+/* Styling the labels */
+label {
+  display: block;
+  margin-bottom: 8px;
+  font-weight: bold;
+  font-size: 14px;
+  color: #2e8b57; /* Dark green for labels */
+}
+
+/* Styling the input fields and select dropdowns */
+input,
+select {
+  width: 100%;
+  padding: 10px;
+  margin-bottom: 15px;
+  border: 1px solid #4caf50; /* Medium green border for inputs */
+  border-radius: 4px;
+  font-size: 14px;
+  box-sizing: border-box;
+}
+
+/* Focus effect for input fields */
+input:focus,
+select:focus {
+  border-color: #388e3c; /* Darker green when focused */
+  outline: none;
+  box-shadow: 0 0 5px rgba(56, 142, 60, 0.5); /* Soft green glow */
+}
+
+/* Submit button styling */
+button {
+  background-color: #388e3c; /* Dark green button */
+  color: white;
+  padding: 12px 20px;
+  border: none;
+  border-radius: 4px;
+  font-size: 16px;
+  cursor: pointer;
+  width: 100%;
+}
+
+/* Hover effect for submit button */
+button:hover {
+  background-color: #2e7d32; /* Darker green for hover effect */
+}
+
+/* Form container styling */
+.fileUpload {
+  padding: 20px;
+  max-width: 800px;
+  margin: 0 auto;
+  background-color: #f1f8f1; /* Light green background for form container */
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+form {
+  padding-top: 20px; /* Add some space from the header */
+}
+
+/* Make the form responsive */
+@media (max-width: 600px) {
+  .fileUpload {
+    padding: 15px;
+  }
+  button {
+    width: auto;
+  }
+}
