+++ react/src/http-common.ts
@@ -1,8 +1,8 @@
 import axios from "axios";
 
 export default axios.create({
-  baseURL: "http://localhost:6000",
-  headers: {
-    "Content-type": "application/json",
-  },
+    baseURL: "",
+    headers: {
+        "Content-type": "application/json",
+    },
 });
