+++ react/src/pages/CSVUpload.tsx
@@ -1,138 +1,60 @@
-import { useState, useEffect } from "react";
-import '../css/StudentForm.css';
-import http from '../http-common.ts';
+import { useState } from "react";
+import '../css/FileUpload.css';
+import UploadService from "../services/FileUploadService.ts";
 import React from "react";
-
-const upload = (file: File, onUploadProgress: any): Promise<any> => {
-    let formData = new FormData();
-
-    formData.append("file", file);
-
-    return http.post("/api/batch_job", formData, {
-        headers: {
-            "Content-Type": "multipart/form-data",
-        },
-        onUploadProgress,
-    });
-};
-
-const getFiles = (): Promise<any> => {
-    return http.get("/files");
-};
-
-const FileUploadService = {
-    upload,
-    getFiles,
-};
-
-interface IFile {
-    url: string,
-    name: string
-}
 
 const FileUpload: React.FC = () => {
 
-
     const [currentFile, setCurrentFile] = useState<File>();
-    const [progress, setProgress] = useState<number>(0);
     const [message, setMessage] = useState<string>("");
-    const [fileInfos, setFileInfos] = useState<Array<IFile>>([])
 
     const selectFile = (event: React.ChangeEvent<HTMLInputElement>) => {
         const { files } = event.target;
         const selectedFiles = files as FileList;
         setCurrentFile(selectedFiles?.[0]);
-        setProgress(0);
     };
 
     const upload = () => {
-        setProgress(0);
         if (!currentFile) return;
 
-        FileUploadService.upload(currentFile, (event: any) => {
-            setProgress(Math.round((100 * event.loaded) / event.total));
-        })
-            .then((response) => {
-                setMessage(response.data.message);
-                return FileUploadService.getFiles();
-            })
-            .then((files) => {
-                setFileInfos(files.data);
-            })
-            .catch((err) => {
-                setProgress(0);
+        UploadService.upload(currentFile, (event: any) => {
+        }).then((response) => {
+            setMessage(response.data.message);
+        }).catch((err) => {
+            console.log(err);
 
-                if (err.response && err.response.data && err.response.data.message) {
-                    setMessage(err.response.data.message);
-                } else {
-                    setMessage("Could not upload the File!");
-                }
+            if (err.response && err.response.data && err.response.data.message) {
+                setMessage(err.response.data.message);
+            } else {
+                setMessage("Could not upload the File!");
+            }
 
-                setCurrentFile(undefined);
-            });
-    };
-
-    useEffect(() => {
-        FileUploadService.getFiles().then((response) => {
-            setFileInfos(response.data);
+            setCurrentFile(undefined);
         });
-    }, []);
+    };
 
 
 
     return (
         <div>
-            <div className="row">
-                <div className="col-8">
-                    <label className="btn btn-default p-0">
-                        <input type="file" onChange={selectFile} />
+            <header className="header">
+                <p>Upload CSV file of multiple students </p>
+            </header>
+            <div className="fileUpload">
+                <div>
+                    <label>
+                        <input type="file" accept=".csv" onChange={selectFile} />
                     </label>
+
+                    <button disabled={!currentFile} onClick={upload}>Upload</button>
                 </div>
 
-                <div className="col-4">
-                    <button
-                        className="btn btn-success btn-sm"
-                        disabled={!currentFile}
-                        onClick={upload}
-                    >
-                        Upload
-                    </button>
-                </div>
-            </div>
-
-            {currentFile && (
-                <div className="progress my-3">
-                    <div
-                        className="progress-bar progress-bar-info"
-                        role="progressbar"
-                        aria-valuenow={progress}
-                        aria-valuemin={0}
-                        aria-valuemax={100}
-                        style={{ width: progress + "%" }}
-                    >
-                        {progress}%
+                {message && (
+                    <div role="alert">
+                        {message}
                     </div>
-                </div>
-            )}
-
-            {message && (
-                <div className="alert alert-secondary mt-3" role="alert">
-                    {message}
-                </div>
-            )}
-
-            <div className="card mt-3">
-                <div className="card-header">List of Files</div>
-                <ul className="list-group list-group-flush">
-                    {fileInfos &&
-                        fileInfos.map((file, index) => (
-                            <li className="list-group-item" key={index}>
-                                <a href={file.url}>{file.name}</a>
-                            </li>
-                        ))}
-                </ul>
-            </div>
-        </div>
+                )}
+            </div></div>
     );
 };
