+++ react/src/services/FileUploadService.ts
@@ -1,0 +1,20 @@
+import http from "../http-common.ts";
+
+const upload = (file: File, onUploadProgress: any): Promise<any> => {
+    let formData = new FormData();
+
+    formData.append("file", file);
+
+    return http.post("/api/batch_job", formData, {
+        headers: {
+            "Content-Type": "multipart/form-data",
+        },
+        onUploadProgress,
+    });
+};
+
+const FileUploadService = {
+    upload,
+};
+
+export default FileUploadService;
